add -k option

+++ ukbb/hesin_release.sh
@@ -13,9 +13,9 @@ function usage () {
     echo ""
     echo "Create a release for inpatient UKBB data"
     echo ""
-    echo "Usage: hesin_release.sh -i <main HESIN table> -d <HESIN DIAG table> -p <HESIN OPER table> -r <release> -x <list of sample IDs to exclude> -o <output dir>"
+    echo "Usage: hesin_release.sh -i <main HESIN table> -d <HESIN DIAG table> -p <HESIN OPER table> -r <release> -o <output dir> { -x <list of sample IDs to exclude> -k <keep temp files>}"
     echo ""
-    echo "Input files and output file are tab-separated."
+    echo "All files are tab-separated."
     echo ""
     exit 0
 }
@@ -27,7 +27,8 @@ fi
 outdir=""
 release=""
 exfile=""
-while getopts "hi:d:x:p:r:o:" opt; do
+keep="NO"
+while getopts "hi:d:x:p:r:o:k" opt; do
     case $opt in
         i)main_fname=($OPTARG);;
         d)diag_fname=($OPTARG);;
@@ -35,6 +36,7 @@ while getopts "hi:d:x:p:r:o:" opt; do
         p)oper_fname=($OPTARG);;
         r)release=($OPTARG);;
         o)outdir=($OPTARG);;
+        k)keep="YES";;
         h)usage;;
         *)usage;;
     esac
@@ -61,6 +63,11 @@ exitIfExists "$outfile" "ERROR: output file $outfile already exists"
 logfile="${outdir}/hesin_r${release}.log"
 outfile=$(realpath "$outfile")
 logfile=$(realpath "$logfile")
+tmpdir=$(mktemp -d -p "$outdir" tempdir_hesin_XXXXXXXX)
+if [[ $? -ne 0 ]];then
+    echo "ERROR: could not create temporary directory in $outdir"
+    exit 1
+fi
 
 : > $logfile
 
@@ -75,13 +82,8 @@ echo "OPER TABLE: $oper_fname" | tee -a "$logfile"
 echo "IDs TO EXCLUDE: $exfile" | tee -a "$logfile"
 echo "OUTPUT DIR: $outdir" | tee -a "$logfile"
 echo "OUTPUT FILE: $outfile" | tee -a "$logfile"
+echo "TEMP DIR: $tmpdir" | tee -a "$logfile"
 echo "" | tee -a "$logfile"
-
-tmpdir=$(mktemp -d -p "$outdir" tempdir_hesin_XXXXXXXX)
-if [[ $? -ne 0 ]];then
-    echo "ERROR: could not create temporary directory in $outdir "|tee -a "$logfile"
-    exit 1
-fi
 
 datestr=$(date +%d-%b-%Y)
 echo $release > "$tmpdir"/RELEASE
@@ -161,6 +163,8 @@ echo "" | tee -a "$logfile"
 echo "CREATING OUTPUT FILE" | tee -a "$logfile"
 cd "$tmpdir" && tar -zcf "$outfile" hesin.txt hesin_diag.txt hesin_oper.txt RELEASE CREATED && cd -
 
-rm -rf "$tmpdir"
+if [[ "$keep" == "NO" ]];then
+    rm -rf "$tmpdir"
+fi
 echo "" | tee -a "$logfile"
 date "+%F %H-%M-%S"|tee -a "$logfile"
